Update database api to use ZODB
Config file is now a specific JSON api implementation for necessary application settings.
Database API now split into 3 classes, one for the database of accounts, accounts themselves, and the monetary splits within the accounts, which are the columns to display.
Also converted the ui files to python files for the upcoming GUI update.