Changes to overwrite/add account entry and table clicking
Added an option to toggle between overwriting an account value or adding to an account value.
Also changed the table itemSelectionChanged signal to cellClicked signal for accuracy as there was an issue where you could change the account in the combo box and try to click on a cell you previously selected and the combo box wouldn't be set properly since selection didn't change.