Minor changes for usability
Added a try catch around the app so that the SystemExit exception is caught and dusted

Abstracted the graph lookback period to a variable to make it easier to change in the future. Also lowered it to 14 from 16 as the date labels were running into each other and becoming unreadable.